refactor(engine): extract TASK.md structural readers to taskdoc.py (engine-modularization 15/N)
Move the 11 TASK.md structural-reader functions out of add.py into a new
add_engine/taskdoc.py module:

  _task_header · _count_test_defs · _primary_test_files · _tests_count
  _declared_test_files · _declared_tests_count · _tests_info · _task_prose
  _phase_spans · _raw_phase_bodies · _spec_delta_entries   (add.py 5891 -> 5680)

A closed, unpatched cluster (transitive-closure AST = zero outbound; the three
add.X test references are plain calls, not patches), so a plain re-export works.
Deps: stdlib (re, Path) + _confined (imported from the components module) + 3
shared delta regexes.

CONSTANT-BY-KIND (shared): _DELTA_RE, _EVIDENCE_RE, and _SPEC_DELTA_RE are used
by BOTH the moved readers (_task_prose / _spec_delta_entries) AND the staying
deltas-web lint (_lint_task_deltas / _collect_open_spec_deltas /
_resolve_spec_delta), so they relocate to constants.py as the single source.
taskdoc.py imports them directly; add.py imports them via the explicit
_-prefixed constants import (so the staying lint still resolves them). Only the
3 regex Assign nodes moved -- the interleaved siblings _SPEC_STATUSES /
_STATUS_SETS / _TAG_BROAD_RE stay (deltas-web only), moved by precise AST
lineno ranges, not text-region deletion.

Provably verbatim: an AST source-segment diff of all 11 functions AND all 3
regexes vs the pre-move add.py is identical. taskdoc.py -> {constants, components}
+ stdlib one-way (imports standalone; no cycle).

test_delta_grammar_dedup.py followed the canonical: its single-source DRY guard
scanned only add.py for the enumerated grammar; since _DELTA_RE legitimately
moved to constants.py, _add_source() was widened to _engine_source() (add.py +
add_engine/*.py). The == 1 invariant is UNCHANGED -- a duplicate anywhere in the
engine still fails -- only the search domain matches the new home. Not a
weakening; the same class as repointing a moved patch target.

Two-pin model re-aimed (both literals; engine_pin.py never hashes):
  ENGINE_MD5     5fb053eb -> fb3ecdcb
  ENGINE_PKG_MD5 fa51a345 -> 4ff845c3

3-tree mirror byte-identical. New test_engine_extract_taskdoc.py (identity ·
shared-regex relocation both sides · siblings kept · no-cycle · 3-tree pins).
Full suite 1948/0; seam audit clean (101).

The engine package is now 13 modules (constants · io_state · accessors ·
predicates · identity · guidelines · render · milestones · components · version ·
release · taskdoc · the entry); add.py 7049 -> 5680 lines across the 15
extractions.

author: Tin Dang
